Function monolithic_userboot::fs::fs_init

pub fn fs_init()

在执行系统调用前初始化文件系统

包括建立软连接，提前准备好一系列的文件与文件夹

Fat32 filesystem doesn’t exists the concept of soft link, so we need to call this function every time we boot the system